Make fontforge completly optional (don't break if you don't have fontforge)
This make fontforge optional which is nice for a community project. It
avoid simple contributors the need to install fontforge (which can
seems heavy, just to update a html file for example).
Also, icons.* font files are now put into src folder so you can commit
these files.
To avoid that behavior, the simplest thing to do is to add them in
.gitignore :)

Close #115